% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{cpg_annotate_chromosome}
\alias{cpg_annotate_chromosome}
\title{Annotate cpgs with chromosome information}
\usage{
cpg_annotate_chromosome(data, var, array_type = c("450K", "EPIC"))
}
\arguments{
\item{data}{A data.frame.}

\item{var}{variable name containing the cpgs.}

\item{array_type}{Type of array annotation data to use. Must be one of
"450K" or "EPIC", defaults to "450K".}
}
\value{
\link[tibble:tibble-package]{tibble} with two columns.
}
\description{
Annotate cpgs with chromosome information
}
\examples{
cpg_annotate_chromosome(tibble::tibble(cpg = sample_cpgs), cpg)
}
