% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{cpg_annotate_region}
\alias{cpg_annotate_region}
\title{Annotate cpg with predefined regions}
\usage{
cpg_annotate_region(
  data,
  var,
  region = c("relation_to_island", "1stExon", "3UTR", "5UTR", "Body", "TSS1500",
    "TSS200"),
  array_type = c("450K", "EPIC")
)
}
\arguments{
\item{data}{A data.frame.}

\item{var}{variable name containing the cpgs.}

\item{region}{character, region to annotate according to. Must be one of
"relation_to_island", "1stExon",  "3UTR", "5UTR", "Body", "TSS1500" or
"TSS200". Defaults to "relation_to_island".}

\item{array_type}{Type of array annotation data to use. Must be one of
"450K" or "EPIC", defaults to "450K".}
}
\value{
\link[tibble:tibble-package]{tibble} with two columns.
}
\description{
Annotate cpg with predefined regions
}
\examples{
df_cpg <- data.frame(cpg = sample_cpgs)
cpg_annotate_region(df_cpg, cpg)
cpg_annotate_region(df_cpg, cpg, "1stExon")
cpg_annotate_region(df_cpg, cpg, "3UTR")
cpg_annotate_region(df_cpg, cpg, "5UTR")
cpg_annotate_region(df_cpg, cpg, "Body")
cpg_annotate_region(df_cpg, cpg, "TSS1500")
cpg_annotate_region(df_cpg, cpg, "TSS200")
}
