% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpg_dist.R
\name{cpg_dist_chromosome}
\alias{cpg_dist_chromosome}
\title{Calculate distribution of cpg over chromosomes}
\usage{
cpg_dist_chromosome(x, array_type = c("450K", "EPIC"), exclude = NULL)
}
\arguments{
\item{x}{vector of cpgs}

\item{array_type}{Type of array annotation data to use. Must be one of
"450K" or "EPIC", defaults to "450K".}

\item{exclude}{Vector of chromosomes to be excludes. Must be prefixed with
"chr".}
}
\value{
\link[tibble:tibble-package]{tibble} with two columns "chr" and "n".
}
\description{
Calculate distribution of cpg over chromosomes
}
\examples{
cpg_dist_chromosome(sample_cpgs)

# Exclude sex chromosomes
cpg_dist_chromosome(sample_cpgs, exclude = c("chrX", "chrY"))
}
