% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_array.R
\name{detect_array}
\alias{detect_array}
\title{Detect array type based on cpgs}
\usage{
detect_array(x)
}
\arguments{
\item{x}{vector of cpgs}
}
\value{
"450k" or "EPIC"
}
\description{
The functions looks to see if any of the cpgs appear in the list of cpgs only
included in 450K, if it finds any it determines 450K, otherwise defaults to
EPIC.
}
\examples{
detect_array(sample_cpgs)
}
