% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{cpg_annotate_chromosome}
\alias{cpg_annotate_chromosome}
\title{Calculate distribution of cpg over chromosomes}
\usage{
cpg_annotate_chromosome(data, var, array_type = c("450K", "EPIC"))
}
\arguments{
\item{array_type}{Type of array annotation data to use. Must be one of
"450K" or "EPIC", defaults to "450K".}

\item{x}{vector of cpgs}

\item{exclude}{Vector of chromosomes to be excludes. Must be prefixed with
"chr".}
}
\value{
\link[tibble:tibble-package]{tibble} with two columns "chr" and "n".
}
\description{
Calculate distribution of cpg over chromosomes
}
\examples{
cpg_annotate_chromosome(tibble::tibble(cpg = sample_cpgs), cpg)
}
